refactor: 문자열 계산기의 전반적인 테스트 코드 수정
