+++ src/test/kotlin/calculator/NumberTest.kt
@@ -1,3 +1,5 @@
+package calculator
+
 import calculator.Number
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
@@ -5,24 +7,24 @@ import io.kotest.inspectors.forAll
 
 class NumberTest : BehaviorSpec({
 
-    val negativeNum = "-1"
     Given("음수가") {
+        val num = "-1"
         When("전달되면") {
             Then("RuntimeException 예외가 발생한다.") {
                 shouldThrow<RuntimeException> {
-                    Number.of(negativeNum)
+                    Number.of(num)
                 }
             }
         }
     }
 
-    val notNum = listOf("*", "!", "a")
     Given("숫자 이외의 값이") {
+        val nums = listOf("*", "!", "a")
         When("전달되면") {
             Then("RuntimeException 예외가 발생한다.") {
-                notNum.forAll {
+                nums.forAll { num ->
                     shouldThrow<RuntimeException> {
-                        Number.of(it)
+                        Number.of(num)
                     }
                 }
             }
